requirements.txt
awsmfunc==1.3.1
beautifulsoup4==4.11.1
cryptography==37.0.2
Custom_Tooltip==1.3
IMDbPY==2022.7.9
Pillow==9.2.0
PyInstaller==5.2
pymediainfo==5.1.0
pyperclip==1.8.2
requests==2.27.1
tkinterdnd2==0.3.0
torf==3.1.3
urllib3
bencode.py
flatbencode
tmdb-python
cffi
cinemagoer
VapourSynth==59
